crud operations by roles
crud operations by roles

+++ src/repository/photos.py
@@ -1,3 +1,5 @@
+from typing import Optional
+
 from fastapi import HTTPException
 from sqlalchemy import select, func
 
@@ -9,7 +11,7 @@ sys.path.insert(0, str(Path(__file__).resolve().parent.parent.parent))
 
 
 from src.schemas.photos import PhotoSchema, PhotoUpdateSchema
-from src.database.models import Photos, User, TransformedPhoto, PhotoTags, TagAssociation
+from src.database.models import Photos, User, TransformedPhoto, PhotoTags, TagAssociation, Role
 from sqlalchemy.ext.asyncio import AsyncSession
 import cloudinary.uploader
 from src.database.models import Photos
@@ -23,7 +25,7 @@ UPLOAD_FOLDER = "uploads"
 
 
 
-async def upload_photo(file_path: str, user_id: int, db: AsyncSession):
+async def upload_photo(file_path: str, user_id: int, db: AsyncSession, user: User):
     cloudinary.config(
         cloud_name=config.cloud_name,
         api_key=config.api_key,
@@ -32,14 +34,18 @@ async def upload_photo(file_path: str, user_id: int, db: AsyncSession):
     response = cloudinary.uploader.upload(file_path)
     photo_url = response['url']
     photo = Photos(photo_link=photo_url, user_id=user_id)
+    if (user.role != Role.admin and photo.user_id != user.id):
+        raise HTTPException(status_code=403, detail="Доступ запрещен")
     db.add(photo)
     await db.commit()
 
     return photo
 async def update_photo(photo_id: int,body: PhotoUpdateSchema, db: AsyncSession, user: User) -> Photos:
-    getting = select(Photos).filter_by(id=photo_id, user=user)
+    getting = select(Photos).filter(Photos.id == photo_id)
     result = await db.execute(getting)
     photo = result.scalar_one_or_none()
+    if (user.role != Role.admin and photo.user_id != user.id):
+        raise HTTPException(status_code=403, detail="Доступ запрещен")
     if photo:
         photo.description = body.description
         await db.commit()
@@ -48,9 +54,11 @@ async def update_photo(photo_id: int,body: PhotoUpdateSchema, db: AsyncSession, 
 
 
 async def delete_photo(photo_id: int, db: AsyncSession, user: User) -> None:
-    getting = select(Photos).filter_by(id=photo_id, user=user)
+    getting = select(Photos).filter(Photos.id == photo_id)
     result = await db.execute(getting)
     photo = result.scalar_one_or_none()
+    if (user.role != Role.admin and photo.user_id != user.id):
+        raise HTTPException(status_code=403, detail="Доступ запрещен")
     if photo:
         await db.delete(photo)
         await db.commit()
@@ -62,7 +70,7 @@ async def get_photo_by_url(url: str, db: AsyncSession, user: User):
     return photo.scalars().first()
 
 
-async def resize_photo(url: str,  width: int, height: int, db: AsyncSession, user: User):
+async def resize_photo(url: str, width: int, height: int, db: AsyncSession, user: User):
     cloudinary.config(
         cloud_name=config.cloud_name,
         api_key=config.api_key,
@@ -70,6 +78,9 @@ async def resize_photo(url: str,  width: int, height: int, db: AsyncSession, use
     )
     response = cloudinary.uploader.upload(url, width=width, height=height)
     resized_url = response['url']
+    photo = await get_photo_by_url(url, db, user)
+    if (user.role != Role.admin and photo.user_id != user.id):
+        raise HTTPException(status_code=403, detail="Доступ запрещен")
     transformed_photo = TransformedPhoto(user_id=user.id, photo_url=resized_url)
     db.add(transformed_photo)
     await db.commit()
@@ -83,6 +94,9 @@ async def crop_photo(url: str,  width: int, height: int, db: AsyncSession, user:
     )
     response = cloudinary.uploader.upload(url, width=width, height=height, crop="crop")
     cropped_url = response['url']
+    photo = await get_photo_by_url(url, db, user)
+    if (user.role != Role.admin and photo.user_id != user.id):
+        raise HTTPException(status_code=403, detail="Доступ запрещен")
     transformed_photo = TransformedPhoto(user_id=user.id, photo_url=cropped_url)
     db.add(transformed_photo)
     await db.commit()
@@ -99,6 +113,9 @@ async def apply_filter_and_text(url: str, filter_name: str, text: str, font_fami
         {"overlay": {"font_family": font_family, "font_size": font_size, "text": text, "font_color": font_color}}
     ])
     edited_url = response['url']
+    photo = await get_photo_by_url(url, db, user)
+    if (user.role != Role.admin and photo.user_id != user.id):
+        raise HTTPException(status_code=403, detail="Доступ запрещен")
     transformed_photo = TransformedPhoto(user_id=user.id, photo_url=edited_url)
     db.add(transformed_photo)
     await db.commit()
@@ -112,6 +129,9 @@ async def convert_photo_format(url: str, new_format: str, db: AsyncSession, user
     )
     response = cloudinary.uploader.upload(url, format=new_format)
     converted_url = response['url']
+    photo = await get_photo_by_url(url, db, user)
+    if (user.role != Role.admin and photo.user_id != user.id):
+        raise HTTPException(status_code=403, detail="Доступ запрещен")
     transformed_photo = TransformedPhoto(user_id=user.id, photo_url=converted_url)
     db.add(transformed_photo)
     await db.commit()
@@ -125,6 +145,9 @@ async def get_photo_metadata(url: str, db: AsyncSession, user: User):
         api_secret=config.api_secret
     )
     response = cloudinary.uploader.upload(url)
+    photo = await get_photo_by_url(url, db, user)
+    if (user.role != Role.admin and photo.user_id != user.id):
+        raise HTTPException(status_code=403, detail="Доступ запрещен")
     metadata = {
         "format": response['format'],
         "width": response['width'],
@@ -142,6 +165,8 @@ async def generate_transformed_photo_link(transformed_photo_id: int, db: AsyncSe
 
 async def generate_qr_code(transformed_photo_id: int, db: AsyncSession, user: User):
     transformed_image = await db.get(TransformedPhoto, transformed_photo_id)
+    if (user.role != Role.admin and transformed_image.user_id != user.id):
+        raise HTTPException(status_code=403, detail="Доступ запрещен")
     if not transformed_image:
         raise HTTPException(status_code=404, detail="Transformed image not found")
     qr = qrcode.QRCode(
@@ -164,6 +189,8 @@ async def generate_qr_code(transformed_photo_id: int, db: AsyncSession, user: Us
 async def create_tag(tag_name: str, db: AsyncSession, user: User):
     existing_tag = await db.execute(select(PhotoTags).filter_by(tag_name=tag_name))
     existing_tag = existing_tag.scalar_one_or_none()
+    if (user.role != Role.admin and existing_tag.user_id != user.id):
+        raise HTTPException(status_code=403, detail="Доступ запрещен")
     if not existing_tag:
         tag = PhotoTags(tag_name=tag_name)
         db.add(tag)
@@ -174,6 +201,8 @@ async def create_tag(tag_name: str, db: AsyncSession, user: User):
 
 async def attach_tag_to_photo(photo_id: int, tag_name: str, db: AsyncSession, user: User):
     photo = await db.get(Photos, photo_id)
+    if (user.role != Role.admin and photo.user_id != user.id):
+        raise HTTPException(status_code=403, detail="Доступ запрещен")
     if not photo:
         raise HTTPException(status_code=404, detail="Photo not found")
 
@@ -191,8 +220,6 @@ async def attach_tag_to_photo(photo_id: int, tag_name: str, db: AsyncSession, us
 
     if count_tags >= 5:
         raise HTTPException(status_code=400, detail="Maximum number of tags reached for this photo")
-
-    # Создаем связь между фотографией и тегом
     tag_association = TagAssociation(photo_id=photo_id, tag_id=tag.tag_id)
 
     db.add(tag_association)
@@ -210,6 +237,9 @@ async def get_tags_for_photo(photo_id: int, db: AsyncSession, user: User):
         select(Photos).filter(Photos.id == photo_id).options(joinedload(Photos.photo_tags))
     )
     photo = photo.scalars().first()
+    if (user.role != Role.admin and photo.user_id != user.id):
+        raise HTTPException(status_code=403, detail="Доступ запрещен")
     if not photo:
         raise HTTPException(status_code=404, detail="Photo not found")
     return photo.photo_tags
+
